refactor: rename interactive to slider
