+++ src/lib/components/hue.ts
@@ -1,4 +1,4 @@
-import { Interactive, Interaction } from './interactive.js';
+import { Slider, Interaction } from './slider.js';
 import { hsvaToHslString } from '../utils/convert.js';
 import { createTemplate } from '../utils/dom.js';
 import { clamp, round } from '../utils/math.js';
@@ -9,7 +9,7 @@ const template = createTemplate(`
 <div role="slider" part="hue" aria-label="Hue" aria-valuemin="0" aria-valuemax="360"><div part="hue-pointer"></div></div>
 `);
 
-export class Hue extends Interactive {
+export class Hue extends Slider {
   private _h!: number;
 
   get xy(): boolean {
